Refactored spectrum averaging into its own class.
