+++ InstrumentControl.sln
@@ -7,7 +7,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InstrumentControl", "Instru
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj", "{2AA5EE2D-A29C-4004-A582-5928B5FE691F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IMSScanClassExtensions", "IMSScanClassExtensions\IMSScanClassExtensions.csproj", "{07B31947-11C8-4B9A-B8CC-813FA9C8AB76}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IMSScanClassExtensions", "IMSScanClassExtensions\IMSScanClassExtensions.csproj", "{07B31947-11C8-4B9A-B8CC-813FA9C8AB76}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpectrumAveraging", "SpectrumAveraging\SpectrumAveraging.csproj", "{0F304F7B-50FF-4314-810A-BEB72140EE58}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{07B31947-11C8-4B9A-B8CC-813FA9C8AB76}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{07B31947-11C8-4B9A-B8CC-813FA9C8AB76}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{07B31947-11C8-4B9A-B8CC-813FA9C8AB76}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0F304F7B-50FF-4314-810A-BEB72140EE58}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0F304F7B-50FF-4314-810A-BEB72140EE58}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0F304F7B-50FF-4314-810A-BEB72140EE58}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0F304F7B-50FF-4314-810A-BEB72140EE58}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
