+++ SpectrumAveraging/RejectionType.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SpectrumAveraging
+{
+    public enum RejectionType
+    {
+        NoRejection,
+        MinMaxClipping,
+        PercentileClipping,
+        SigmaClipping,
+        WinsorizedSigmaClipping,
+        AveragedSigmaClipping,
+        BelowThresholdRejection
+    }
+
+
+}
